Fix unintended branching.
JGitAdapter contained an inconsistency when checking the need for
branching: Previously an 'outgoing edge count' of 1 or higher was
sufficient to trigger branching. But since the pending workflow step has
already been added to the workflow graph, this would always be
triggered. Now a count of 2 or greater is needed.