Replace current IdentifiableResolver implementation with a no-op one.
The currently used IdentifiableResolver is misbehaving in that it
incorrectly prioritizes the checksum attribute of files too high and as
such fails in almost all situations involving empty or copied files.

As such we have replaced it by a no-op version that never tries to
actually resolve a file, but only creates a simple resource descriptor
out of its identifiers.